refactor: svelte actions
