+++ web/src/lib/actions/focus.ts
@@ -1,0 +1,3 @@
+export const initInput = (element: HTMLInputElement) => {
+  element.focus();
+};
