+++ web/src/lib/actions/shortcut.ts
